\name{add.map.grid}
\alias{add.map.grid}
\title{ Geographic grid }
\description{
  Adds geographic longitude / latitude circles (grid-lines) on a pre-existing
  plot generated by plot.map. Grid-lines should be added immediately
  after producing the underlying projection. 
}
\usage{
add.map.grid(lon.grds = seq(-180, 180, by = 10), 
             lat.grds = seq(-80, +80, by = 10), 
             delta.draw = 45/200, col = "black", lwd = 0.5, lty = 3)
}
\arguments{
  \item{lon.grds}{ Vector of longitude values for which latitude circles
      shall be plotted. }
  \item{lat.grds}{ Vector of latitude values for which latitude circles
      shall be plotted.  }
  \item{delta.draw}{ Number of points to use for drawing a 45 degree
      segment of a longitude or latitude circle. }
  \item{col}{ Color of the circles. }
  \item{lwd}{ Line width of the circles. }
  \item{lty}{ Line type of the circles }
}
\details{
  
}
\value{
  Only plotting, no data output.
}
\references{  }
\author{ Christoph Frei }
\note{ 
}
\seealso{ \code{\link{add.map.outlines}} }
\examples{
}
\keyword{ hplot }
