\name{add.map.outlines}
\alias{add.map.outlines}
\title{ Map outlines }
\description{
  Adds geographic map outlines on a pre-existing plot generated by plot.map. 
  Map outlines can be national boundaries, rivers and lakes or continent
  outlines. Map outlines should be added to a plot immediately after
  producing the underlying projection.
}
\usage{
add.map.outlines(mapdat, lonlim.map, latlim.map, 
                 thin.map = 0, thin.map2 = thin.map,
                 col = "black", lwd = 1, col2 = col, lwd2 = lwd)
}
\arguments{
  \item{mapdat}{ Description of a map outlines dataset. (See details.)
    If \code{mapdat="none"} no map outlines are drawn. }
  \item{lonlim.map}{ Range of longitude values for which to look for
    map outlines to be plotted. A two-element vector. }
  \item{latlim.map}{ Range of latitude values for which to look for
    map outlines to be plotted. A two-element vector. }
  \item{thin.map}{ Number of iterations to thin out national boundaries
            or continents specified by \code{mapdat}. It is useful when
            \code{mapdat="worldHires"} or \code{mapdat="continents"}. If 
            \code{thin.map>0} the line segments for the map outline are 
            diluted. \code{thin.map = 1} means that every second point 
            is omitted. \code{thin.map = 2} means that this thinning 
            out is applied twice, etc. Thinning is useful to prevent 
            excessively large Postscript files just because of high-
            resolution national boundaries or continents.
            \code{thin.map = 0} (the default) means no thinning. }
  \item{thin.map2}{ The same like \code{thin.map} but for rivers and lakes. }
  \item{col}{ Colour of map outlines (borders). }
  \item{lwd}{ Line width of map outlines (borders). }
  \item{col2}{ Colour of map outlines (rivers and lakes). }
  \item{lwd2}{ Line width of map outlines (rivers and lakes). }
}
\details{
  ** Available map outlines datasets **

  Several datasets of map outlines can be used:

  \code{mapdat="world"} is a coarse resolution of national boundaries,

  \code{mapdat="worldHires"} is a high resolution national boundaries,

  \code{mapdat="continents"} is continental outlines only,

  \code{mapdat="swiss:border&lakes&rivers"} is outlines for Switzerland,

  \code{mapdat="alps:borders&lakes&rivers"} is outlines for the Alpine 
  region. 

  For the last two, this includes boundaries, lakes and rivers. 
  The sub-elements can be chosen as you desire.

  Except for "continents", "swiss:??" and "alps:??" the map outlines are 
  taken from
  packages \pkg{maps} and \pkg{mapdata}. Actually these datasets can be
  accessed very flexibly. For more information check see 
  \code{\link[maps]{map}}. 


  ** Nice to know **

  This function is executed with every call to \code{plot.map}. Then
  colour und line width cannot be modified there. User-defined plotting of
  map outlines is possible by calling \code{plot.map} with 
  \code{mapdat="none"} and a subsequent "manual" call of 
  \code{add.map.outlines}. 
}
\value{
  No output, just plotting.
}
\references{  }
\author{ Christoph Frei and Francesco Isotta }
\note{ 
}
\seealso{ \code{\link{plot.map}} }
\examples{
}
\keyword{ hplot }
