\name{cmsaf.cols}
\alias{cmsaf.cols}
\docType{data}
\title{ Color table from cmsaf }
\description{
  A three column data.frame with r,g,b values of colors. This dataset
  is not meant to be used by the user directly. The cmsaf color table
  can be accessed via \code{cmsaf.col()}.
}
\usage{data(cmsaf.cols)}
\format{
  A data frame with 256 observations on the following 3 variables.
  \describe{
    \item{\code{r}}{red values}
    \item{\code{g}}{green values}
    \item{\code{b}}{blue values}
  }
}
\details{
  
}
\source{
  Table provided by Bruno Duerr
}
\references{
  
}
\examples{
data(cmsaf.cols)
}
\keyword{datasets}
